Move consumers.py into a consumers module.
Break consumers out into base consumers, and each new implmentation of a 
consumer can have its own file.
